+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte
 
+# 04/03/2023
+- Ajout bouton, fonction Eteindre
+- Correction de bugs
+
 # 21/10/2022
 - Adaptation commandes pour v4.3